Renamed all test folders from "Test" to lower case "test"
Also got CppUTest building on 32 bit.  Hopefully I didn't break the 64 bit
version :P

+++ Common/CppUTest/include/CppUTest/SimpleString.h
@@ -96,7 +96,7 @@ private:
    SimpleStringCollection(SimpleStringCollection&);
 };
 
-SimpleString StringFrom (size_t value);
+SimpleString StringFrom (unsigned long value);
 SimpleString StringFrom (bool value);
 SimpleString StringFrom (void* value);
 SimpleString StringFrom (char value);
